Move providers includes to common include area
