+++ include/providers/bio.h
@@ -10,7 +10,7 @@
 #include <stdarg.h>
 #include <openssl/bio.h>
 #include <openssl/core.h>
-#include "prov/provider_ctx.h"
+#include <providers/provider_ctx.h>
 
 int ossl_prov_bio_from_dispatch(const OSSL_DISPATCH *fns);
