+++ include/providers/macsignature.h
@@ -10,7 +10,7 @@
 #include <stdlib.h>
 #include <openssl/crypto.h>
 #include "internal/refcount.h"
-#include "prov/provider_util.h"
+#include <providers/provider_util.h>
 
 struct mac_key_st {
     OSSL_LIB_CTX *libctx;
